buildfix on MSYS
