+++ toolchain-powerpc.cmake
@@ -18,8 +18,8 @@ if(NOT LIBOGCDIR)
     set(LIBOGCDIR ${DEVKITPRO}/libogc)
 endif()
 
-set(CMAKE_C_COMPILER ${DEVKITPPC}/bin/powerpc-eabi-gcc)
-set(CMAKE_CXX_COMPILER ${DEVKITPPC}/bin/powerpc-eabi-g++)
+set(CMAKE_C_COMPILER ${DEVKITPPC}/bin/powerpc-eabi-gcc.exe)
+set(CMAKE_CXX_COMPILER ${DEVKITPPC}/bin/powerpc-eabi-g++.exe)
 set(CMAKE_FIND_ROOT_PATH ${DEVKITPPC})
 set(CMAKE_FIND_ROOT_PATH_MODE_PROGRAM NEVER)
 set(CMAKE_FIND_ROOT_PATH_MODE_LIBRARY ONLY)
